+++ src/main/java/in/koreatech/koin/domain/owner/dto/email/OwnerSendEmailRequest.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.owner.dto;
+package in.koreatech.koin.domain.owner.dto.email;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
@@ -8,10 +8,10 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
-public record VerifyEmailRequest(
-    @JsonProperty("address")
+public record OwnerSendEmailRequest(
     @Email(message = "이메일 형식이 올바르지 않습니다. ${validatedValue}")
     @NotBlank(message = "이메일은 필수입니다.")
+    @JsonProperty(value = "address")
     @Schema(description = "이메일", example = "temp@gmail.com", requiredMode = REQUIRED)
     String address
 ) {
